Remove reboot functionality.
Rebooting is best handled outside of the system's login logic. If it's
desired that logging in with a special username triggers a reboot, this
can be set up with a combination of PAM module and user shell. On
systems where this is not desired, having the option may lead to
disruptive misconfigurations.